fix(algorithm): address Sage round 3 — honest doc + gate-policy layering
- HonestOracle: architecture.md now states the gate refuses a `passed` whose
  evidence kind is explicitly `specified`, and that legacy criteria with no
  recorded kind are grandfathered — not an absolute integrity guarantee.
- Layering: move learnGateViolations from isa-accessors (structural access) into
  algorithm.ts (the Algorithm gate module); the pure predicates isClosedCriterion
  / isHollowPass stay in isa-accessors. Gate policy now lives with the gate.

typecheck + lint clean; affected suites green.
