+++ src/algorithm.ts
@@ -19,11 +19,29 @@ import {
   buildIsaArtifact,
   defaultEvidenceKind,
   getCriteria,
-  learnGateViolations,
+  isClosedCriterion,
+  isHollowPass,
   progressFromCriteria,
   updateCriterionWithResult,
   verifiedFromCriteria,
 } from "./isa-accessors";
+
+/**
+ * The criteria that block entry to LEARN, split by reason. Single source of truth
+ * for the Algorithm's LEARN-gate policy — both the assertGate guard and sync's
+ * reachability check call this so the two cannot drift when a future evidence rule
+ * is added. Composes the pure isa-accessor predicates; gate policy lives here in
+ * the Algorithm module, not in the structural accessor layer.
+ */
+export function learnGateViolations(criteria: readonly IdealStateCriterion[]): {
+  unresolved: IdealStateCriterion[];
+  hollow: IdealStateCriterion[];
+} {
+  return {
+    unresolved: criteria.filter((criterion) => !isClosedCriterion(criterion)),
+    hollow: criteria.filter(isHollowPass),
+  };
+}
 import { getRunPhase } from "./algorithm-lifecycle";
 import { DEFAULT_ALGORITHM_LOOP_STATE } from "./algorithm-execution-modes";
 import { appendAlgorithmProvenance } from "./algorithm-provenance";
